refactor(ClientsTable): extract logic to custom hook
