+++ src/features/clients/api/index.ts
@@ -1,0 +1,4 @@
+export * from './deleteClient';
+export * from './updateClient';
+export * from './createClient';
+export * from './getClientsList';
