+++ src/features/clients/components/ClientsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './ClientsTable';
